# api/openapi.yaml
openapi: 3.0.3
info:
  title: Adera API
  version: 1.0.0
  description: |
    Trusted Ethiopian review platform backend (አደራ — "balance").

    ## Conventions
    - All endpoints live under `/api/v1`. Requests and responses are JSON (UTF-8);
      user content may be Amharic, English, Afaan Oromo, Tigrinya, or Somali.
    - Success bodies are `{"data": ..., "meta": {...}?}`; errors are
      `{"error": {"code", "message", "details"?, "request_id"}}` with **stable
      machine-readable codes** for client-side localization.
    - Timestamps are RFC 3339 UTC. Public identifiers are UUIDs.
    - Unknown JSON fields are **rejected** (400 `invalid_json`) as a
      mass-assignment defense. Bodies are capped at 64 KiB (uploads bypass the
      API via presigned POST).
    - Cursor pagination: pass `meta.next_cursor` back as `?cursor=`; cursors are
      opaque. `limit` defaults to 20, max 50.
    - Sensitive endpoints are rate limited (429 + `Retry-After`).

    ## Authentication
    `Authorization: Bearer <access_token>` (15-minute JWT). Refresh tokens are
    opaque, single-use, and rotate on every refresh; reusing a rotated token
    revokes its whole session family. Roles: customer, business_owner
    (granted only by an approved claim), moderator, admin (implies moderator).

    ## Review submission flow
    1. `GET /categories/{c}/criteria` → render category-specific questions.
    2. `POST /reviews` with `Idempotency-Key` header (safe to retry on weak
...

...
servers:
  - url: /
...
security: []
paths:
...
  /api/v1/targets:
    post:
      tags: [targets]
      summary: Submit a review target (business location, online seller, product, ...)
      description: |
        Public submissions enter moderation (`pending`); moderator submissions
        publish immediately. Aliases (Amharic spellings, transliterations)
        power mixed-script search. `social_links` values must be https URLs on
        the platform's own domain (tiktok/instagram/facebook/telegram/youtube).
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TargetCreate'}
      responses:
        '201':
          description: Created (check `moderation_status`).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataEnvelope'
                  - properties:
                      data: {$ref: '#/components/schemas/Target'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {$ref: '#/components/responses/ValidationError'}
    get:
      tags: [targets]
      summary: Browse published targets (filters + cursor pagination, newest first)
      parameters:
        - {$ref: '#/components/parameters/FilterCategory'}
        - {$ref: '#/components/parameters/FilterCity'}
        - {$ref: '#/components/parameters/FilterArea'}
        - {$ref: '#/components/parameters/FilterType'}
        - {$ref: '#/components/parameters/FilterMinRating'}
        - {$ref: '#/components/parameters/FilterVerified'}
        - {name: online_only, in: query, schema: {type: boolean}}
        - {name: business, in: query, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Page of targets.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TargetPage'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/targets/top-rated:
    get:
      tags: [targets]
      summary: Top-rated targets by the documented confidence-adjusted score
      description: |
        Ordered by `ranking_score = (10×3.5 + Σratings)/(10 + n)` — Bayesian
        shrinkage so one 5★ review can't outrank hundreds of consistent
        reviews. The raw average is returned unaltered alongside.
      parameters:
        - {$ref: '#/components/parameters/FilterCategory'}
        - {$ref: '#/components/parameters/FilterCity'}
        - {$ref: '#/components/parameters/FilterArea'}
        - {$ref: '#/components/parameters/FilterType'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Ranked targets (includes `ranking_score`).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RankedTargetPage'}
  /api/v1/targets/trending:
    get:
      tags: [targets]
      summary: Trending targets (published-review volume, last 30 days)
      parameters:
        - {$ref: '#/components/parameters/FilterCategory'}
        - {$ref: '#/components/parameters/FilterCity'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Targets with `recent_review_count` and `recent_average`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TrendingTargetPage'}
  /api/v1/targets/nearby:
    get:
      tags: [targets]
      summary: Published targets within a radius of a point, closest first
      description: |
        Great-circle (haversine) distance against the stored coordinates; a
        bounding box narrows the scan first so the geo index is usable.
        Targets without coordinates (online sellers) are never returned.
        Each item carries `distance_km`. Results are a single ranked page —
        widen `radius_km` or raise `limit` rather than paginating.
      parameters:
        - {name: lat, in: query, required: true, schema: {type: number, minimum: -90, maximum: 90}, example: 9.0}
        - {name: lng, in: query, required: true, schema: {type: number, minimum: -180, maximum: 180}, example: 38.75}
        - name: radius_km
          in: query
          schema: {type: number, minimum: 0, exclusiveMinimum: true, maximum: 50, default: 5}
          description: Search radius in kilometres; at most 50.
        - {$ref: '#/components/parameters/FilterCategory'}
        - {$ref: '#/components/parameters/FilterCity'}
        - {$ref: '#/components/parameters/FilterArea'}
        - {$ref: '#/components/parameters/FilterType'}
        - {$ref: '#/components/parameters/FilterMinRating'}
        - {$ref: '#/components/parameters/FilterVerified'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Nearest targets first (each includes `distance_km`).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NearbyTargetPage'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/targets/{idOrSlug}:
    get:
      tags: [targets]
      summary: Get one target by UUID or slug
      description: Unpublished targets are 404 except for moderators.
      parameters: [{$ref: '#/components/parameters/PathIDOrSlug'}]
      responses:
        '200':
          description: Target with aliases and rating summary.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataEnvelope'
                  - properties:
                      data: {$ref: '#/components/schemas/Target'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [targets]
      summary: Update a claimed target (business members or moderators)
      description: |
        Explicit field allowlist; name/category changes are moderator-only
        (via edit suggestions) to prevent identity swaps.
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/PathIDOrSlug'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TargetUpdate'}
      responses:
        '200':
          description: Updated target.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataEnvelope'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/targets/{idOrSlug}/edit-suggestions:
    post:
      tags: [targets]
      summary: Suggest a correction (any authenticated user; moderators apply)
      security: [{bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/PathIDOrSlug'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [changes]
              properties:
                changes:
                  type: object
                  description: Allowlisted keys only (name, description, address_text, phone, website, city_id, area_id, online_only, social_links, aliases, category_id, target_type).
                note: {type: string, maxLength: 1000}
      responses:
        '201': {$ref: '#/components/responses/StatusCreated'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PathIDOrSlug:
      name: idOrSlug
      in: path
      required: true
      schema: {type: string}
      description: Target UUID or slug.
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
      description: Opaque cursor from a previous page's `meta.next_cursor`.
    Limit:
      name: limit
      in: query
      schema: {type: integer, default: 20, minimum: 1, maximum: 50}
    FilterCategory:
      name: category
      in: query
      schema: {type: string, format: uuid}
    FilterCity:
      name: city
      in: query
      schema: {type: string, format: uuid}
    FilterArea:
      name: area
      in: query
      schema: {type: string, format: uuid}
    FilterType:
      name: type
      in: query
      schema: {$ref: '#/components/schemas/TargetType'}
    FilterMinRating:
      name: min_rating
      in: query
      schema: {type: number, minimum: 1, maximum: 5}
    FilterVerified:
      name: verified
      in: query
      schema: {type: boolean}
      description: Only targets with verified status.
  responses:
...
    StatusCreated:
      description: 'Created: `{"data": {"id": "...", "status": "..."}}`'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/DataEnvelope'}
...
    ValidationError:
      description: Validation failed (code `validation_failed`, per-field `details`).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            error:
              code: validation_failed
              message: invalid review
              details: {overall_rating: must be 1-5}
              request_id: 4f33f9884ec53c39
    Unauthorized:
      description: Missing/invalid credentials (codes `unauthorized`, `invalid_credentials`).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Forbidden:
      description: Authenticated but not allowed (code `forbidden`, or `account_suspended` on login).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    NotFound:
      description: Not found — also used instead of 403 where existence itself is private.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
    DataEnvelope:
      type: object
      properties:
        data: {}
        meta: {$ref: '#/components/schemas/PageMeta'}
    PageMeta:
      type: object
      properties:
        next_cursor: {type: string}
        has_more: {type: boolean}
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine code for localization.
              enum: [validation_failed, invalid_json, unauthorized, invalid_credentials,
                     forbidden, not_found, conflict, gone, rate_limited, payload_too_large,
                     internal_error, account_suspended, cooldown_active, upload_invalid,
                     precondition_failed, verification_unavailable, storage_unavailable, not_ready]
            message: {type: string}
            details:
              type: object
              additionalProperties: {type: string}
            request_id: {type: string}
...
    TargetType:
      type: string
      enum: [business, location, online_seller, product, service, repair_provider, restaurant, cafe]
...
    Target:
      type: object
      properties:
        id: {type: string, format: uuid}
        target_type: {$ref: '#/components/schemas/TargetType'}
        category_id: {type: string, format: uuid}
        business_id: {type: string, format: uuid, nullable: true}
        name: {type: string}
        slug: {type: string, example: tomoca-coffee-piassa-a1b2c3d4}
        description: {type: string}
        city_id: {type: string, format: uuid, nullable: true}
        area_id: {type: string, format: uuid, nullable: true}
        address_text: {type: string, description: Landmark-relative free text ("behind Edna Mall").}
        latitude: {type: number, nullable: true}
        longitude: {type: number, nullable: true}
        online_only: {type: boolean}
        phone: {type: string}
        website: {type: string}
        social_links:
          type: object
          additionalProperties: {type: string}
          example: {telegram: "https://t.me/selamstore"}
        aliases:
          type: array
          items: {type: string}
          example: ["ቶሞካ ቡና", "Tomoca"]
        verification_status: {type: string, enum: [unverified, verified]}
        moderation_status: {type: string, enum: [pending, published, hidden, removed, merged]}
        review_count: {type: integer}
        average_rating: {type: number, nullable: true, description: "Raw mean, 2 decimals, never adjusted."}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    TargetCreate:
      type: object
      required: [target_type, category_id, name]
      properties:
        target_type: {$ref: '#/components/schemas/TargetType'}
        category_id: {type: string, format: uuid}
        business_id: {type: string, format: uuid, nullable: true}
        name: {type: string, minLength: 2, maxLength: 160}
        description: {type: string, maxLength: 2000}
        city_id: {type: string, format: uuid, nullable: true}
        area_id: {type: string, format: uuid, nullable: true}
        address_text: {type: string, maxLength: 300}
        latitude: {type: number, nullable: true}
        longitude: {type: number, nullable: true}
        online_only: {type: boolean}
        phone: {type: string}
        website: {type: string}
        social_links: {type: object, additionalProperties: {type: string}}
        aliases: {type: array, maxItems: 5, items: {type: string}}
    TargetUpdate:
      type: object
      description: Allowlisted fields only.
      properties:
        description: {type: string}
        address_text: {type: string}
        phone: {type: string}
        website: {type: string}
        city_id: {type: string, format: uuid}
        area_id: {type: string, format: uuid}
        online_only: {type: boolean}
        social_links: {type: object, additionalProperties: {type: string}}
    TargetPage:
      allOf:
        - $ref: '#/components/schemas/DataEnvelope'
        - properties:
            data:
              type: array
              items: {$ref: '#/components/schemas/Target'}
    RankedTarget:
      allOf:
        - $ref: '#/components/schemas/Target'
        - properties:
            ranking_score: {type: number, description: "(10×3.5 + Σratings)/(10+n), 4 decimals."}
    RankedTargetPage:
      allOf:
        - $ref: '#/components/schemas/DataEnvelope'
        - properties:
            data:
              type: array
              items: {$ref: '#/components/schemas/RankedTarget'}
    TrendingTarget:
      allOf:
        - $ref: '#/components/schemas/Target'
        - properties:
            recent_review_count: {type: integer, description: "Published reviews in the last 30 days."}
            recent_average: {type: number, nullable: true, description: "Average rating over that same window."}
    TrendingTargetPage:
      allOf:
        - $ref: '#/components/schemas/DataEnvelope'
        - properties:
            data:
              type: array
              items: {$ref: '#/components/schemas/TrendingTarget'}
    NearbyTarget:
      allOf:
        - $ref: '#/components/schemas/Target'
        - properties:
            distance_km: {type: number, description: "Great-circle distance from the query point."}
    NearbyTargetPage:
      allOf:
        - $ref: '#/components/schemas/DataEnvelope'
        - properties:
            data:
              type: array
              items: {$ref: '#/components/schemas/NearbyTarget'}
...